Otorisasi taks

+++ app/Http/Controllers/TaskController.php
@@ -3,9 +3,12 @@
 namespace App\Http\Controllers;
 
 use App\Models\Task;
+
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Redis;
+
 
 class TaskController extends Controller
 {
@@ -37,6 +40,7 @@ class TaskController extends Controller
     {
         $pageTitle = 'Edit Task';
         $task = Task::find($id);
+        Gate::authorize('update', $task);
         session()->put('previousPage', url()->previous());
         return view('tasks.edit', ['pageTitle' => $pageTitle, 'task' => $task]);
     }
@@ -57,7 +61,8 @@ class TaskController extends Controller
             'detail' => $request->detail,
             'due_date' => $request->due_date,
             'status' => $request->status,
-            'user_id' => Auth::user()->id,
+            'user_id' => Auth::user()->id
+
         ]);
         $previousPage = session('previousPage');
 
@@ -66,6 +71,7 @@ class TaskController extends Controller
 
         // Redirect sesuai dengan halaman sebelumnya
         return redirect()->to($previousPage ?: route('tasks.index'));
+
     }
 
     public function update(Request $request, $id)
@@ -80,32 +86,37 @@ class TaskController extends Controller
         );
 
         $task = Task::find($id);
+        Gate::authorize('update', $task);
         $task->update([
             'name' => $request->name,
             'detail' => $request->detail,
             'due_date' => $request->due_date,
             'status' => $request->status,
+            
         ]
     );
     $previousPage = session('previousPage');
 
-    // Bersihkan session
-    session()->forget('previousPage');
 
-    // Redirect sesuai dengan halaman sebelumnya
-    return redirect()->to($previousPage ?: route('tasks.index'));
+        // Bersihkan session
+        session()->forget('previousPage');
+
+        // Redirect sesuai dengan halaman sebelumnya
+        return redirect()->to($previousPage ?: route('tasks.index'));
     }
 
     public function delete($id)
     {
         $pageTitle = 'Delete Task';
         $task = Task::find($id);
+        Gate::authorize('delete', $task);
         return view('tasks.delete', ['pageTitle' => $pageTitle, 'task' => $task]);
     }
 
     public function destroy($id)
     {
         $task = Task::find($id);
+        Gate::authorize('delete', $task);
         $task->delete();
         return redirect()->route('tasks.index');
     }
@@ -115,18 +126,19 @@ class TaskController extends Controller
     {
         $title = 'Task Progress';
         $tasks = Task::all();
+       
         $filteredTasks = $tasks->groupBy('status');
         // $tasks = [
-        //     'not_started' => $filteredTasks->get('not_started' , []),
-        //     'in_progress' => $filteredTasks->get('in_progress', []),
-        //     'completed' => $filteredTasks->get('completed', []),
-        //     'in_review' => $filteredTasks->get('in_review', []),
-        // ];
-        $tasks = [
-            Task::STATUS_NOT_STARTED => $filteredTasks->get(
-                Task::STATUS_NOT_STARTED, []
-            ),
-            Task::STATUS_IN_PROGRESS => $filteredTasks->get(
+            //     'not_started' => $filteredTasks->get('not_started' , []),
+            //     'in_progress' => $filteredTasks->get('in_progress', []),
+            //     'completed' => $filteredTasks->get('completed', []),
+            //     'in_review' => $filteredTasks->get('in_review', []),
+            // ];
+            $tasks = [
+                Task::STATUS_NOT_STARTED => $filteredTasks->get(
+                    Task::STATUS_NOT_STARTED, []
+                ),
+                Task::STATUS_IN_PROGRESS => $filteredTasks->get(
                 Task::STATUS_IN_PROGRESS, []
             ),
             Task::STATUS_IN_REVIEW => $filteredTasks->get(
@@ -146,7 +158,7 @@ class TaskController extends Controller
     public function move(int $id, Request $request)
     {
     $task = Task::findOrFail($id);
-
+    Gate::authorize('update', $task);
     $task->update([
         'status' => $request->status,
     ]);
@@ -164,5 +176,23 @@ class TaskController extends Controller
      {
         return back()->withInput();
      }
+    }
+
+    public function home()
+    {
+    $tasks = Task::where('user_id', auth()->id())->get();
+
+    $completed_count = $tasks
+        ->where('status', Task::STATUS_COMPLETED)
+        ->count();
+
+    $uncompleted_count = $tasks
+        ->whereNotIn('status', Task::STATUS_COMPLETED)
+        ->count();
+
+    return view('home', [
+        'completed_count' => $completed_count,
+        'uncompleted_count' => $uncompleted_count,
+    ]);
     }
 }
